requirements.txt
# lib
numpy
scipy
torch==1.11.0
torchvision==0.12.0
pillow
pandas
scikit-learn
tqdm
tensorboard
timm @ git+https://github.com/rwightman/pytorch-image-models
pytorch-metric-learning
hydra-core==1.1
hydra_colorlog==1.1.0

# utils
matplotlib
SciencePlots
plotext
jupyter
ipython

# dev
ipdb
flake8
flake8-annotations
autopep8
pylint
isort
pytest
